add column to db

+++ pkg/db/migrations/000035_add_inclusion_delay_column.up.sql
@@ -1,1 +1,1 @@
-ALTER TABLE t_validator_rewards_summary ADD COLUMN f_inclusion_delay INT NOT NULL DEFAULT 0;
+ALTER TABLE t_validator_rewards_summary ADD COLUMN f_inclusion_delay SMALLINT NOT NULL DEFAULT 0;
